Add :readmore and :readmore-material modules
