+++ sample/build.gradle
@@ -43,6 +43,7 @@ android {
 dependencies {
     implementation project(':material')
     implementation project(':visibility')
+    implementation project(':readmore-material')
 
     implementation libs.androidx.activity.compose
     implementation libs.compose.foundation.foundation
