+++ sample/src/main/java/soup/metronome/sample/readmore/ReadMoreTextDemo.kt
@@ -36,6 +36,7 @@ import androidx.compose.ui.text.style.TextDecoration
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import soup.metronome.readmore.material.ReadMoreText
 import soup.metronome.sample.theme.SampleTheme
 
 private const val description: String =
