+++ settings.gradle
@@ -2,3 +2,5 @@ include ':sample'
 include ':material'
 include ':visibility'
 include ':internal-testutils'
+include ':readmore'
+include ':readmore-material'
